+++ analysis/frontend/src/components/google/google-map.css
@@ -1,0 +1,2 @@
+.gm-style-cc { display:none; }
+
