+++ analysis/frontend/src/index.tsx
@@ -1,21 +1,19 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import './styles/index.css';
-import App from './components/App-google';
+import App from './components/google/App-google';
 // import App from './components/App-gaode';
 // import App from './App-deckGL.js';
 import reportWebVitals from './reportWebVitals';
 
 const root = ReactDOM.createRoot(
-  document.getElementById('root') as HTMLElement
+    document.getElementById('root') as HTMLElement
 );
 
 
 root.render(
-  //  https://stackoverflow.com/a/72371185/9422455
-  <React.Fragment>
-    <App />
-  </React.Fragment>
+    //  https://stackoverflow.com/a/72371185/9422455
+    <App/>
 );
 
 // If you want to start measuring performance in your app, pass a function
